+++ dev_env/src/README.md
@@ -1,0 +1,1 @@
+Agent which implements the Hyperspace Neighbor Penetration algorithm developed [here](https://arxiv.org/ftp/arxiv/papers/2106/2106.05497.pdf)
