+++ dev_env/src/hnp.egg-info/not-zip-safe
@@ -1,0 +1,1 @@
+
