+++ dev_env/src/hnp/hnp.egg-info/PKG-INFO
@@ -1,0 +1,10 @@
+Metadata-Version: 1.0
+Name: hnp
+Version: 0.1
+Summary: Hyperspace Neighbor Penetration agent
+Home-page: #
+Author: Thomas Jiralerspong
+Author-email: thomasjiralerspong@gmail.com
+License: UNKNOWN
+Description: UNKNOWN
+Platform: UNKNOWN
