+++ dev_env/src/hnp/hnp.egg-info/SOURCES.txt
@@ -1,0 +1,6 @@
+setup.py
+hnp.egg-info/PKG-INFO
+hnp.egg-info/SOURCES.txt
+hnp.egg-info/dependency_links.txt
+hnp.egg-info/not-zip-safe
+hnp.egg-info/top_level.txt
