+++ dev_env/src/hnp/hnp.egg-info/not-zip-safe
@@ -1,0 +1,1 @@
+
