+++ dev_env/src/pyproiect.toml
@@ -1,0 +1,19 @@
+name = "hnp"
+version = "0.0.1"
+authors = [
+  { name="Thomas Jiralerspong", email="thomasjiralerspong@gmail.com" },
+  { name="John Willes", email="john.willes@vectorinstitute.ai" },
+  { name="Marshall Wang", email="marshall.wang@vectorinstitute.ai" },
+]
+description = "Agent which implements the Hyperspace Neighbour Penetration algorithm"
+readme = "README.md"
+requires-python = ">=3.8"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+]
+
+[project.urls]
+"Homepage" = "https://github.com/VectorInstitute/MBRL-HVAC-Energy-Optimization"
+"Bug Tracker" = "https://github.com/VectorInstitute/MBRL-HVAC-Energy-Optimization/issues"
