+++ dev_env/src/setup.py
@@ -1,0 +1,11 @@
+import setuptools
+
+setuptools.setup(name='hnp',
+version='0.1',
+description='Hyperspace Neighbor Penetration agent',
+url='#',
+author='Thomas Jiralerspong',
+install_requires=[],
+author_email='thomasjiralerspong@gmail.com',
+packages=setuptools.find_packages(),
+zip_safe=False)
